django/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
neo4j==5.26.0
